began overhaul
This commit marks the beginning of a serious overhaul of this readout
software with the intent on making it simple, extendable, and far more
reliable. A core function required to meet base requirements is the
ability to sweep an LO source while taking data. This is the current
focus for the following set of commits.