added paramsweep and udptest
param sweep needs work to improve the baseband target frequencies. I've realized
that I can write a function to import the data from a copper mountain VNA and
use that data to generate sweep targets. it's less of a feature and more of a nice to have...

udptest needs to be tested to ensure it works properly